refactor(observability): trim tracing surface
